Add files via upload
Hey guys, dunno if this is the proper way to use github, but here is the base stuff so far. What needs work is the reportNum() method in TILE. It should be reporting the number of mines touching that particular object. I think we might want to move it to the BOARDMAKER class which is in charge of the actual grid. Let me know what you think