Add files via upload
Okay, this is really basic but I can't take it the rest of the way cuz i'm not familiar enough with the gui stuff. 
Two methods getHighScores() and createHighScores()
if you just wanna look at the high scores use getHighScores() it returns the contents of the highscores.txt file as a ArrayList of strings called scores. If it can't find a file it'll create one which we may have to take out. I'm not sure that's the correct thing todo.

The create highscores doesn't work yet because you have to type in the path of where you want it to go. I dunno how to fix this yet. Also the variable score that's written to the file needs to be read from the timer value when the game is over. THEN it needs to be sorted based on lowest score. I dunno if we can do it in time